refactor(server): user create logic
